Añadir proveedor de tema y eliminar código innecesario en la página de inicio de sesión
