+++ src/main.jsx
@@ -14,13 +14,17 @@ import 'react-quill/dist/quill.snow.css';
 import { ReactQueryDevtools } from '@tanstack/react-query-devtools'
 // Supports weights 100-900
 import '@fontsource-variable/montserrat';
+import { ThemeProvider } from "./context/provider-theme.jsx";
 const queryClient = new QueryClient();
 
 ReactDOM.createRoot(document.getElementById("root")).render(
   <React.StrictMode>
     <BrowserRouter>
       <QueryClientProvider client={queryClient}>
+        <ThemeProvider>
         <App /> <Toaster expand={false} closeButton />
+        </ThemeProvider>
+        
         <ReactQueryDevtools initialIsOpen={false} />
       </QueryClientProvider>
     </BrowserRouter>
