+++ src/pages/LoginPag/Login.jsx
@@ -13,24 +13,6 @@ import axiosInstance from "@/services/ConfigApi";
 import { UsecontextAuth } from "@/context/provider-auth";
 
 function Login() {
-  const [ThemeActual, setThemeActual] = useState(() => {
-    if (window.matchMedia("(prefers-color-scheme: dark)").matches) {
-      return "dark";
-    } else {
-      return "light";
-    }
-  });
-  useEffect(() => {
-    if (ThemeActual == "dark") {
-      document.querySelector("html").classList.add("dark");
-    } else {
-      document.querySelector("html").classList.remove("dark");
-    }
-  }, [ThemeActual]);
-
-  function HandleTheme() {
-    setThemeActual((value) => (value == "light" ? "dark" : "light"));
-  }
   const [showPassword, setShowPassword] = useState(false);
 
   const handlePasswordVisibility = () => {
@@ -64,7 +46,8 @@ function Login() {
   return (
     <main className=" overflow-x-hidden h-screen grid place-content-center dark:bg-DarkFondo  dark:text-white relative">
       <div className="absolute top-5 right-5">
-        <button onClick={HandleTheme}>
+        {/**
+           * <button onClick={HandleTheme}>
           {" "}
           {ThemeActual === "light" ? (
             <IconMoonFilled />
@@ -72,6 +55,8 @@ function Login() {
             <IconBrightnessUp />
           )}{" "}
         </button>
+           * 
+           */}
       </div>
       <article className=" py-4   lg:py-4 px-10 lg:px-8 bg-white  dark:bg-[#484848] lg:w-[620px] lg:shadow-2xl rounded-xl ">
         <section>
